Making my header "work" as intended
I had an issue where if I was on the checkout page, both the cart icon and the "shop" navlink would have the "active" class.

I get that this is due to them being in the same route, but the navlinks are not given the same class in the header component. I don't fully get why that happens, but I have fixed the issue by forcing an "inactive" class which overwrites the "active" class when I am in the checkout page.

Not a pretty solution and I don't fully understand why this happens, but I need to move on to bigger things in the project and not muck about in the header