Weird issue
Having a strange issue where I can set the topResult variable, and console log it just fine. And then I can set the title, etc, and it works just fine

But when the website is reloaded, I get an error where "topResult" is undefined. No idea how to fix it, but I need a break now