Many changes
Key notes:

I've separated the data -> component part of the featured item component to another file to keep things clean.

Had an issue where the topResult was not set before destructuring began, probably because it is dependent on an async get function. Got issues where the destrucutred data was undefined, but fixed it by adding "=topResult || {}" (or empty object).