Tidying up component
The featured item is coming along now, but is still missing a lot of content. Firstly, it's not as pretty as I would like, and secondly, I need to get on the buttons components.

Also a lot of cleanup to do in the "App.Styles.Js" file, which I should get rid of in favor of component files. Will get on it tomorrow