Setting up a loader and squishing bugs
Loader animation is now ok, but it only lasts a short time. If I find time to prioritize this, I'll delay the get function so that the loader has more purpose

Also fixed an issue where the a p element was a child of another p element. Not how it's supposed to be