More setup for the search function
I'm getting there, at the moment I can log info from the search input, but clearly that is not enough.

I'm working on how I want my code to look, and thinking about how to handle sorting and searching moving forward.

I will most likely handle the sorting and searching of the array by saving it in a state, and then passing it through whatever function I need to. And for that I will probably need state management, and I will probably go for Zustand