Search function "works"
I keep getting weird errors when trying to use useState in my search function. I'm sure that this has something to do with how my component is set up, but I can't seem to figure it out.

At the moment, I have a search function that does search through my array and gives me what I want. But I need to be able to use this data in a meaningful way.  Not sure how yet