Installing Zustand and messing around
I installed Zustand for state management, but it's not quite time to use it yet. I'll make use of it later for the shopping cart, but it's not needed for the search function

I made usestate work after all on my search function, but lifting up state from child components. Not really elegant, but it works