More work on sort and search
I have an issue where I have to set the displayed array to the orignal array when there is no search query.

But this means that the webshop is no longer sorted. gotta figure out how to remedy this.

A fix might be to have another setState, which I call "setSorted" or something, which then is passed to the searchbar. I'll figure it out