Still not happy with this
I can sort and I can search, but I can't search, sort and then remove the search term.

This is because I set the sorted array based on the output when I change the sorting. And the output was set to the search-results.

If I change it so that the sort is done with the original array, then the search becomes invalid after I change the sorting.

I'm sure there's an easy fix, but I can't see it yet